в дебагер добавленна возможность дизассемблирования кода

+++ OVM_SDK/OVM_SDK.h
@@ -30,8 +30,8 @@ enum class Registers{
 };
 
 struct SectorDescription{
-    int start;
-    int size;
+    long long start;
+    long long size;
 };
 
 inline std::pair<bool, std::pair<int, int>> decodeAddress(long long address){
